# apps/inngest-app/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inngest App API",
    "version": "1.0.0",
    "description": "API documentation for the Inngest app"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format: Bearer {token}"
      }
    },
...
  },
  "paths": {
...
    "/api/workflows": {
      "get": {
        "summary": "Get user's workflows",
        "description": "Returns a list of all workflows for the authenticated user.",
        "tags": [
          "Workflows"
        ],
        "responses": {
          "200": {
            "description": "List of user workflows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "available": {
                            "type": "boolean"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "IDLE",
                              "RUNNING",
                              "COMPLETED",
                              "FAILED",
                              "CANCELLED"
                            ]
                          },
                          "canBeScheduled": {
                            "type": "boolean"
                          },
                          "eventName": {
                            "type": "string"
                          },
                          "cronExpressions": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "timezone": {
                            "type": "string"
                          },
                          "lastRunAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "nextRunAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "input": {
                            "type": "object"
                          },
                          "requiredProviders": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "credentials": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "type": {
                                  "type": "string"
                                },
                                "provider": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}